fix: update timestamp in header and change ft_puthex parameter type to unsigned long long

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lmelo-do <lmelo-do@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/04 19:29:02 by lmelo-do          #+#    #+#             */
-/*   Updated: 2025/08/04 19:29:04 by lmelo-do         ###   ########.fr       */
+/*   Updated: 2025/08/07 20:01:22 by lmelo-do         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_putchar(char c);
 int	ft_putstr(const char *s);
 int	ft_putnbr(int n);
 int	ft_putnbr_unsigned(unsigned int n);
-int	ft_puthex(unsigned int n, int uppercase);
+int	ft_puthex(unsigned long long n, int uppercase);
 int	ft_putptr(void *ptr);
 
 #endif
